+++ docs/changelog.md
@@ -2,6 +2,10 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format. 
 
+## [Unreleased]
+### Added
+- Ability to control behavior when pulled asset already exists
+
 ## [0.1.2] - 2019-06-11
 ### Fixed
 - Logic path such that genome config file is updated regardless of asset unpacking